# core/test/api/contracts/v1/contacts.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ONECore API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/contacts/by-codes": {
      "get": {
        "summary": "Get multiple contacts by their contact codes",
        "description": "Fetch a batch of contacts by providing a comma-separated list of contact codes.",
        "tags": ["Contacts"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "anyOf": [
                          {
                            "type": "object",
                            "properties": {
                              "contactCode": {
                                "type": "string"
                              },
                              "communication": {
                                "type": "object",
                                "properties": {
                                  "phoneNumbers": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "phoneNumber": {
                                          "type": "string"
                                        },
                                        "type": {
                                          "type": "string",
                                          "enum": [
                                            "work",
                                            "home",
                                            "mobile",
                                            "direct-line",
                                            "fax",
                                            "pager",
                                            "unspecified"
                                          ]
                                        },
                                        "comment": {
                                          "type": "string"
                                        },
                                        "isPrimary": {
                                          "type": "boolean"
                                        }
                                      },
                                      "required": [
                                        "phoneNumber",
                                        "type",
                                        "isPrimary"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "emailAddresses": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "emailAddress": {
                                          "type": "string"
                                        },
                                        "type": {
                                          "type": "string"
                                        },
                                        "isPrimary": {
                                          "type": "boolean"
                                        }
                                      },
                                      "required": [
                                        "emailAddress",
                                        "type",
                                        "isPrimary"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "specialAttention": {
                                    "type": "boolean"
                                  }
                                },
                                "required": [
                                  "phoneNumbers",
                                  "emailAddresses",
                                  "specialAttention"
                                ],
                                "additionalProperties": false
                              },
                              "addresses": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "careOf": {
                                      "type": "string"
                                    },
                                    "street": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "zipCode": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "city": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "region": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "country": {
                                      "type": "string",
                                      "nullable": true
                                    }
                                  },
                                  "required": [
                                    "street",
                                    "zipCode",
                                    "city",
                                    "region",
                                    "country"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "type": {
                                "type": "string",
                                "enum": ["individual"]
                              },
                              "personal": {
                                "type": "object",
                                "properties": {
                                  "nationalRegistrationNumber": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "birthDate": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "firstName": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "lastName": {
                                    "type": "string",
                                    "nullable": true
                                  },
                                  "fullName": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "nationalRegistrationNumber",
                                  "birthDate",
                                  "firstName",
                                  "lastName",
                                  "fullName"
                                ],
                                "additionalProperties": false
                              },
                              "trustee": {
                                "type": "object",
                                "properties": {
                                  "contactCode": {
                                    "type": "string"
                                  },
                                  "fullName": {
                                    "type": "string"
                                  }
                                },
                                "required": ["contactCode"],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "contactCode",
                              "communication",
                              "addresses",
                              "type",
                              "personal"
                            ],
                            "additionalProperties": false
                          },
                          {
                            "type": "object",
                            "properties": {
                              "contactCode": {
                                "type": "string"
                              },
                              "communication": {
                                "type": "object",
                                "properties": {
                                  "phoneNumbers": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "phoneNumber": {
                                          "type": "string"
                                        },
                                        "type": {
                                          "type": "string",
                                          "enum": [
                                            "work",
                                            "home",
                                            "mobile",
                                            "direct-line",
                                            "fax",
                                            "pager",
                                            "unspecified"
                                          ]
                                        },
                                        "comment": {
                                          "type": "string"
                                        },
                                        "isPrimary": {
                                          "type": "boolean"
                                        }
                                      },
                                      "required": [
                                        "phoneNumber",
                                        "type",
                                        "isPrimary"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "emailAddresses": {
                                    "type": "array",
                                    "items": {
                                      "type": "object",
                                      "properties": {
                                        "emailAddress": {
                                          "type": "string"
                                        },
                                        "type": {
                                          "type": "string"
                                        },
                                        "isPrimary": {
                                          "type": "boolean"
                                        }
                                      },
                                      "required": [
                                        "emailAddress",
                                        "type",
                                        "isPrimary"
                                      ],
                                      "additionalProperties": false
                                    }
                                  },
                                  "specialAttention": {
                                    "type": "boolean"
                                  }
                                },
                                "required": [
                                  "phoneNumbers",
                                  "emailAddresses",
                                  "specialAttention"
                                ],
                                "additionalProperties": false
                              },
                              "addresses": {
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "careOf": {
                                      "type": "string"
                                    },
                                    "street": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "zipCode": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "city": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "region": {
                                      "type": "string",
                                      "nullable": true
                                    },
                                    "country": {
                                      "type": "string",
                                      "nullable": true
                                    }
                                  },
                                  "required": [
                                    "street",
                                    "zipCode",
                                    "city",
                                    "region",
                                    "country"
                                  ],
                                  "additionalProperties": false
                                }
                              },
                              "type": {
                                "type": "string",
                                "enum": ["organisation"]
                              },
                              "organisation": {
                                "type": "object",
                                "properties": {
                                  "organisationNumber": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  }
                                },
                                "required": ["organisationNumber", "name"],
                                "additionalProperties": false
                              }
                            },
                            "required": [
                              "contactCode",
                              "communication",
                              "addresses",
                              "type",
                              "organisation"
                            ],
                            "additionalProperties": false
                          }
                        ]
                      }
                    }
                  },
                  "required": ["content"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_links": {}
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codes",
            "description": "Comma-separated list of contact codes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
...
}